// WYNKWeb/src/app/Models/Investigationimage.model.ts
export class InvestigationImages {

  CmpID = 0;
  ID: number;
  UIN = '';
  RegistrationTranID = 0;
  InvestigationID: number;
  InvestigationDescription = '';
  ImageLocation = '';
  ExternalInternal: string = null;
  ImageCapturedLocation: string = null;
  InvestigationAmount = 0;
  Remarks: string = null;
  Address1: string = null;
  Address2: string = null;
  ReferredBy: string = null;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  CreatedBy: number;
  UpdatedBy = 0;
  TaxID = 0;
  TaxPercentage = 0;
  TaxValue = 0;
  DiscountPercentage = 0;
  DiscountValue = 0;
  TotalValue = 0;

}